move delete handler to provider
